+++ src/Controller/PublisherController.php
@@ -10,13 +10,12 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
-
 class PublisherController extends AbstractController
 {
     #[Route('/publisher/index', name: 'app_publisher')]
     public function index(): Response
     {
-        $this->denyAccessUnlessGranted('IS_AUTHENTICATED_FULLY', null, 'Not Allowed Access');
+        $this->denyAccessUnlessGranted('ROLE_USER', null, 'Not Allowed Access');
 
         return $this->render('publisher/index.html.twig', [
         ]);
@@ -25,6 +24,7 @@ class PublisherController extends AbstractController
 	#[Route(path: '/publisher/ajax', name: 'app_publisher_ajax', methods: ['POST'])]
     public function ajax(DataTableService $dataTable, EntityManagerInterface $entityManager, Request $request) : Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER', null, 'Not Allowed Access');
         $dataTable->setColumnOrder([null, null, 'name', 'address']);
         $dataTable->setColumnSearch(['id', 'name', 'address']);
         $dataTable->setTable('publisher');
@@ -54,6 +54,7 @@ class PublisherController extends AbstractController
 	#[Route('/publisher/{id}/show', name: 'app_publisher_show', methods: ['GET'])]
     public function show(Publisher $publisher): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER', null, 'Not Allowed Access');
         return $this->render('publisher/show.html.twig', [
             'publisher' => $publisher,
         ]);
@@ -62,9 +63,9 @@ class PublisherController extends AbstractController
 	#[Route(path: '/publisher/new', name: 'app_publisher_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, ): Response
     {
-        $publisher = new Publisher();
+        $this->denyAccessUnlessGranted('ROLE_USER', null, 'Not Allowed Access');
 
-        
+        $publisher = new Publisher();
 
         $form = $this->createForm(PublisherType::class, $publisher);
         $form->handleRequest($request);
@@ -84,14 +85,13 @@ class PublisherController extends AbstractController
 	#[Route('/publisher/{id}/edit', name: 'app_publisher_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Publisher $publisher, int $id, EntityManagerInterface $entityManager, ): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER', null, 'Not Allowed Access');
+
         if (!$publisher) {
             throw $this->createNotFoundException(
                 'No Publisher found for id '.$id
             );
         }
-
-        
-
 
         $form = $this->createForm(publishertype::class, $publisher);
         $form->handleRequest($request);
@@ -112,6 +112,8 @@ class PublisherController extends AbstractController
 	#[Route('/publisher/{id}/delete', name: 'app_publisher_delete', methods: ['POST'])]
     public function delete(EntityManagerInterface $entityManager, Request $request, Publisher $publisher): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER', null, 'Not Allowed Access');
+        
         if ($this->isCsrfTokenValid('delete'.$publisher->getId(), $request->request->get('_token'))) {
             $entityManager->remove($publisher);
             $entityManager->flush();
